+++ indra-core/src/main/java/org/lambda3/indra/core/VectorSpaceFactory.java
@@ -1,5 +1,7 @@
 package org.lambda3.indra.core;
 
+import org.lambda3.indra.core.composition.VectorComposerFactory;
+
 public abstract class VectorSpaceFactory<T extends VectorSpace> extends IndraCachedFactory<T> {
     protected VectorComposerFactory vectorComposerFactory;
