+++ indra-core/src/main/java/org/lambda3/indra/core/composition/VectorComposerFactory.java
@@ -1,8 +1,11 @@
-package org.lambda3.indra.core;
+package org.lambda3.indra.core.composition;
 
-import org.lambda3.indra.client.ScoreFunction;
+
+import org.lambda3.indra.core.composition.*;
 import org.lambda3.indra.core.exception.IndraError;
-import org.lambda3.indra.core.impl.*;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /*-
  * ==========================License-Start=============================
@@ -16,10 +19,10 @@ import org.lambda3.indra.core.impl.*;
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -29,35 +32,23 @@ import org.lambda3.indra.core.impl.*;
  * THE SOFTWARE.
  * ==========================License-End===============================
  */
-public class RelatednessFunctionFactory {
 
-    public RelatednessFunction create(ScoreFunction scoreFunction) {
+public class VectorComposerFactory {
 
-        switch (scoreFunction) {
-            case COSINE:
-                return new CosineRelatednessFunction();
-            case SPEARMAN:
-                return new SpearmanRelatednessFunction();
-            case EUCLIDEAN:
-                return new EuclideanRelatednessFunction();
-            case JACCARD:
-                return new JaccardRelatednessFunction();
-            case PEARSON:
-                return new PearsonRelatednessFunction();
-            case ALPHASKEW:
-                return new AlphaSkewRelatednessFunction();
-            case CHEBYSHEV:
-                return new ChebyshevRelatednessFunction();
-            case CITYBLOCK:
-                return new CityBlockRelatednessFunction();
-            case DICE:
-                return new DiceRelatednessFunction();
-            case JACCARD2:
-                return new Jaccard2RelatednessFunction();
-            case JENSENSHANNON:
-                return new JensenShannonRelatednessFunction();
-            default:
-                throw new IndraError("Unsupported Score Function.");
+    private Map<VectorComposition, VectorComposer> statelessComposers = new HashMap<>();
+
+    public VectorComposerFactory() {
+        this.statelessComposers.put(VectorComposition.SUM, new SumVectorComposer());
+        this.statelessComposers.put(VectorComposition.UNIQUE_SUM, new UniqueSumVectorComposer());
+        this.statelessComposers.put(VectorComposition.AVERAGE, new AveragedVectorComposer());
+    }
+
+    public VectorComposer getComposer(VectorComposition name) {
+        VectorComposer composer = statelessComposers.get(name);
+        if (composer == null) {
+            throw new IndraError("Unsupported vector composition.");
         }
+
+        return composer;
     }
 }
