+++ indra-core/src/main/java/org/lambda3/indra/core/function/ChebyshevRelatednessFunction.java
@@ -1,4 +1,4 @@
-package org.lambda3.indra.core.impl;
+package org.lambda3.indra.core.function;
 
 /*-
  * ==========================License-Start=============================
@@ -26,10 +26,10 @@ package org.lambda3.indra.core.impl;
  * ==========================License-End===============================
  */
 
-import org.apache.commons.math3.linear.RealVector;
-import org.lambda3.indra.core.RelatednessFunction;
 
-public class Jaccard2RelatednessFunction implements RelatednessFunction {
+import org.apache.commons.math3.linear.RealVector;
+
+public class ChebyshevRelatednessFunction implements RelatednessFunction {
 
     @Override
     public double sim(RealVector r1, RealVector r2, boolean sparse) {
@@ -37,23 +37,15 @@ public class Jaccard2RelatednessFunction implements RelatednessFunction {
             return 0;
         }
 
-        double min = 0.0;
-        double max = 0.0;
+        double max = 0;
+        double tmp;
 
-        for (int i = 0; i <r1.getDimension(); ++i) {
-            if (r1.getEntry(i) > r2.getEntry(i)) {
-                min +=r2.getEntry(i);
-                max += r1.getEntry(i);
-            } else {
-                min += r1.getEntry(i);
-                max += r2.getEntry(i);
-            }
+        for (int i = 0; i < r1.getDimension(); ++i) {
+            tmp = Math.abs((r1.getEntry(i) - r2.getEntry(i)));
+            max = (tmp > max ? tmp : max);
         }
 
-        if (max == 0) {
-            return 0;
-        }
-
-        return Math.abs(min / max);
+        double result = 1 / (1 + (max == Double.NaN ? 0 : max));
+        return Math.abs(result);
     }
 }
