+++ indra-core/src/main/java/org/lambda3/indra/core/function/CosineRelatednessFunction.java
@@ -1,4 +1,4 @@
-package org.lambda3.indra.core.impl;
+package org.lambda3.indra.core.function;
 
 /*-
  * ==========================License-Start=============================
@@ -27,13 +27,11 @@ package org.lambda3.indra.core.impl;
  */
 
 import org.apache.commons.math3.linear.RealVector;
-import org.lambda3.indra.core.RelatednessFunction;
-import org.lambda3.indra.core.exception.IndraError;
 
-public class JaccardRelatednessFunction implements RelatednessFunction {
+public class CosineRelatednessFunction implements RelatednessFunction {
 
     @Override
     public double sim(RealVector r1, RealVector r2, boolean sparse) {
-        throw new IndraError("Score function not implemented yet!");
+        return r1.cosine(r2);
     }
 }
