+++ indra-core/src/main/java/org/lambda3/indra/core/function/RelatednessFunction.java
@@ -1,4 +1,4 @@
-package org.lambda3.indra.core.impl;
+package org.lambda3.indra.core.function;
 
 /*-
  * ==========================License-Start=============================
@@ -12,10 +12,10 @@ package org.lambda3.indra.core.impl;
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -26,27 +26,9 @@ package org.lambda3.indra.core.impl;
  * ==========================License-End===============================
  */
 
-
 import org.apache.commons.math3.linear.RealVector;
-import org.lambda3.indra.core.RelatednessFunction;
 
-public class ChebyshevRelatednessFunction implements RelatednessFunction {
+public interface RelatednessFunction {
 
-    @Override
-    public double sim(RealVector r1, RealVector r2, boolean sparse) {
-        if (r1.getDimension() != r2.getDimension()) {
-            return 0;
-        }
-
-        double max = 0;
-        double tmp;
-
-        for (int i = 0; i < r1.getDimension(); ++i) {
-            tmp = Math.abs((r1.getEntry(i) - r2.getEntry(i)));
-            max = (tmp > max ? tmp : max);
-        }
-
-        double result = 1 / (1 + (max == Double.NaN ? 0 : max));
-        return Math.abs(result);
-    }
+    double sim(RealVector r1, RealVector r2, boolean sparse);
 }
